seperate post and get from main
