+++ internal/handler/url_test.go
@@ -1,4 +1,4 @@
-package main
+package handler
 
 import (
 	"net/http"
